Add online test with some coverage
