+++ .gitignore
@@ -1,1 +1,4 @@
 node_modules/
+.env
+.DS_Store
+coverage/
