+++ package.json
@@ -1,10 +1,13 @@
 {
-  "name": "testjs",
-  "version": "1.0.0",
-  "description": "",
-  "main": "test.js",
+  "name": "libra-service",
+  "version": "1.1.0",
+  "description": "Libra Service",
+  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test-dev": "jest --watchAll",
+    "coverage": "jest --coverage",
+    "dev": "nodemon index.js",
     "start": "node index.js"
   },
   "keywords": [],
@@ -18,5 +21,10 @@
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
     "shelljs": "^0.8.3"
+  },
+  "devDependencies": {
+    "jest": "^24.8.0",
+    "nodemon": "^1.19.1",
+    "supertest": "^4.0.2"
   }
 }
